tests/requirements.txt
python-sonarqube-api >= 2.0
requests >= 2.31
